Add Distance Calculation to Airport Aggregate

+++ Tests/MicroLine.Services.Airline.Tests.Unit/Domain/Airports/AirportLocationTests.cs
@@ -19,7 +19,7 @@ public class AirportLocationTests
     {
         // Given
         // When
-        var action = () => AirportLocation.Create(Continent.NorthAmerica, country, "Ontario", "Toronto", 53.309700012200004m, -113.580001831m);
+        var action = () => AirportLocation.Create(Continent.NorthAmerica, country, "Ontario", "Toronto", 53.309700012200004, -113.580001831);
 
         // Then
         action.Should().ThrowExactly<InvalidAirportLocationException>()
@@ -32,7 +32,7 @@ public class AirportLocationTests
     {
         // Given
         // When
-        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", region, "Toronto", 53.309700012200004m, -113.580001831m);
+        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", region, "Toronto", 53.309700012200004, -113.580001831);
 
         // Then
         action.Should().ThrowExactly<InvalidAirportLocationException>()
@@ -45,7 +45,7 @@ public class AirportLocationTests
     {
         // Given
         // When
-        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", city, 53.309700012200004m, -113.580001831m);
+        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", city, 53.309700012200004, -113.580001831);
 
         // Then
         action.Should().ThrowExactly<InvalidAirportLocationException>()
@@ -53,20 +53,20 @@ public class AirportLocationTests
     }
 
 
-    public static TheoryData<decimal> InvalidLatitudeList = new()
+    public static TheoryData<double> InvalidLatitudeList = new()
     {
-        -91.309700012200004m,
-        91.309700012200004m,
-        91.0000000000m,
-        -91.0000000000m
+        -91.309700012200004,
+        91.309700012200004,
+        91.0000000000,
+        -91.0000000000
     };
 
     [Theory, MemberData(nameof(InvalidLatitudeList))]
-    public void AirportLocation_ShouldThrowInvalidAirportLocationException_WhenItCreatesAndLatitudeIsInvalid(decimal latitude)
+    public void AirportLocation_ShouldThrowInvalidAirportLocationException_WhenItCreatesAndLatitudeIsInvalid(double latitude)
     {
         // Given
         // When
-        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", "Toronto", latitude, -113.580001831m);
+        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", "Toronto", latitude, -113.580001831);
 
         // Then
         action.Should().ThrowExactly<InvalidAirportLocationException>()
@@ -74,20 +74,20 @@ public class AirportLocationTests
     }
 
 
-    public static TheoryData<decimal> InvalidLongitudeList = new()
+    public static TheoryData<double> InvalidLongitudeList = new()
     {
-        -181.580001831m,
-        181.580001831m,
-        185.000001m,
-        -198.012001m,
+        -181.580001831,
+        181.580001831,
+        185.000001,
+        -198.012001,
     };
 
     [Theory, MemberData(nameof(InvalidLongitudeList))]
-    public void AirportLocation_ShouldThrowInvalidAirportLocationException_WhenItCreatesAndLongitudeIsInvalid(decimal longitude)
+    public void AirportLocation_ShouldThrowInvalidAirportLocationException_WhenItCreatesAndLongitudeIsInvalid(double longitude)
     {
         // Given
         // When
-        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", "Toronto", 53.309700012200004m, longitude);
+        var action = () => AirportLocation.Create(Continent.NorthAmerica, "Canada", "Ontario", "Toronto", 53.309700012200004, longitude);
 
         // Then
         action.Should().ThrowExactly<InvalidAirportLocationException>()
@@ -103,8 +103,8 @@ public class AirportLocationTests
         var country = "Canada";
         var region = "Ontario";
         var city = "Toronto";
-        var latitude = 53.309700012200004m;
-        var longitude = -113.580001831m;
+        var latitude = 53.309700012200004;
+        var longitude = -113.580001831;
 
         var expected = $"{country}, {region}, {city}, [{latitude}, {longitude}]";
